services/feature_dashboard/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["shahnoor77 <shahnoorkhan9955@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
streamlit = "^1.35.0"
hopsworks = "^3.7.0"
pydantic-settings = "^2.3.3"
loguru = "^0.7.2"
bokeh = "2.4.3"
python-dotenv = "^1.1.0"
hsfs = "3.7.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
